+++ tests/tests_data/streaming/test_serializer.py
@@ -26,7 +26,7 @@ from lightning.data.streaming.serializers import (
     _SERIALIZERS,
     _TORCH_DTYPES_MAPPING,
     _TORCH_VISION_AVAILABLE,
-    IntSerializer,
+    IntegerSerializer,
     JPEGSerializer,
     NoHeaderNumpySerializer,
     NoHeaderTensorSerializer,
@@ -42,24 +42,14 @@ _PIL_AVAILABLE = RequirementCache("PIL")
 
 
 def test_serializers():
-    assert list(_SERIALIZERS.keys()) == [
-        "video",
-        "tif",
-        "file",
-        "pil",
-        "int",
-        "jpeg",
-        "bytes",
-        "no_header_numpy",
-        "numpy",
-        "no_header_tensor",
-        "tensor",
-        "pickle",
-    ]
+    keys = list(_SERIALIZERS.keys())
+    assert keys == [
+        'str', 'int', 'float', 'video', 'tif', 'file', 'pil', 'jpeg',
+        'bytes', 'no_header_numpy', 'numpy', 'no_header_tensor', 'tensor', 'pickle']
 
 
 def test_int_serializer():
-    serializer = IntSerializer()
+    serializer = IntegerSerializer()
 
     for i in range(100):
         data, _ = serializer.serialize(i)